+++ app/src/main/java/woowacourse/shopping/data/dto/response/ProductDto.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.dto
+package woowacourse.shopping.data.dto.response
 
 import com.example.domain.model.Price
 import com.google.gson.annotations.SerializedName
@@ -9,6 +9,7 @@ data class ProductDto(
 
     @SerializedName("name")
     val name: String,
+
     @SerializedName("imageUrl")
     val imgUrl: String,
