+++ app/src/main/java/woowacourse/shopping/mapper/ProductDaoMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.mapper
 
 import com.example.domain.model.Product
-import woowacourse.shopping.data.dto.ProductDto
+import woowacourse.shopping.data.dto.response.ProductDto
 
 fun ProductDto.toDomain(): Product =
     Product(id, name, imgUrl, price)
